move classes + add docs
