+++ kernel/kernel-defaults/src/test/scala/io/delta/kernel/defaults/CommitIcebergActionSuite.scala
@@ -23,6 +23,7 @@ import scala.collection.immutable.Seq
 
 import io.delta.kernel.{Table, Transaction}
 import io.delta.kernel.data.Row
+import io.delta.kernel.defaults.utils.WriteUtils
 import io.delta.kernel.engine.Engine
 import io.delta.kernel.exceptions.KernelException
 import io.delta.kernel.internal.{TableConfig, TableImpl}
