+++ kernel/kernel-defaults/src/test/scala/io/delta/kernel/defaults/utils/WriteUtils.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.delta.kernel.defaults
+package io.delta.kernel.defaults.utils
 
 import java.io.File
 import java.nio.file.{Files, Paths}
@@ -24,22 +24,19 @@ import scala.collection.JavaConverters._
 import scala.collection.immutable.{ListMap, Seq}
 
 import io.delta.golden.GoldenTableUtils.goldenTablePath
-import io.delta.kernel.{Meta, Operation, Table, Transaction, TransactionBuilder, TransactionCommitResult}
-import io.delta.kernel.Operation.MANUAL_UPDATE
+import io.delta.kernel._
 import io.delta.kernel.data.{ColumnarBatch, ColumnVector, FilteredColumnarBatch, Row}
 import io.delta.kernel.defaults.internal.data.DefaultColumnarBatch
-import io.delta.kernel.defaults.utils.{TestRow, TestUtils}
 import io.delta.kernel.engine.Engine
 import io.delta.kernel.expressions.{Column, Literal}
 import io.delta.kernel.expressions.Literal.ofInt
 import io.delta.kernel.hook.PostCommitHook.PostCommitHookType
-import io.delta.kernel.internal.{ScanImpl, SnapshotImpl, TableConfig, TableImpl, TransactionImpl}
+import io.delta.kernel.internal._
 import io.delta.kernel.internal.actions.{DomainMetadata, Metadata, Protocol, SingleAction}
 import io.delta.kernel.internal.fs.{Path => DeltaPath}
-import io.delta.kernel.internal.util.{Clock, FileNames, VectorUtils}
+import io.delta.kernel.internal.util.{Clock, FileNames}
 import io.delta.kernel.internal.util.SchemaUtils.casePreservingPartitionColNames
-import io.delta.kernel.internal.util.Utils.singletonCloseableIterator
-import io.delta.kernel.internal.util.Utils.toCloseableIterator
+import io.delta.kernel.internal.util.Utils.{singletonCloseableIterator, toCloseableIterator}
 import io.delta.kernel.statistics.DataFileStatistics
 import io.delta.kernel.types.IntegerType.INTEGER
 import io.delta.kernel.types.StructType
@@ -52,101 +49,16 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import org.apache.commons.io.FileUtils
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.Path
-import org.scalatest.funsuite.AnyFunSuite
 
-trait TransactionBuilderSupport {
-
-  def createWriteTxnBuilder(
-      table: Table,
-      operation: Operation = Operation.WRITE): TransactionBuilder
-
-  // scalastyle:off argcount
-  def createTxn(
-      engine: Engine,
-      tablePath: String,
-      isNewTable: Boolean = false,
-      schema: StructType = null,
-      partCols: Seq[String] = null,
-      tableProperties: Map[String, String] = null,
-      clock: Clock = () => System.currentTimeMillis,
-      withDomainMetadataSupported: Boolean = false,
-      maxRetries: Int = -1,
-      clusteringColsOpt: Option[List[Column]] = None,
-      logCompactionInterval: Int = 10,
-      txnId: Option[(String, Long)] = None,
-      tablePropertiesRemoved: Set[String] = null): Transaction
-  // scalastyle:on argcount
-}
-
-trait TransactionBuilderV1Support extends TransactionBuilderSupport with TestUtils {
-
-  // scalastyle:off argcount
-  override def createTxn(
-      engine: Engine,
-      tablePath: String,
-      isNewTable: Boolean = false,
-      schema: StructType = null,
-      partCols: Seq[String] = null,
-      tableProperties: Map[String, String] = null,
-      clock: Clock = () => System.currentTimeMillis,
-      withDomainMetadataSupported: Boolean = false,
-      maxRetries: Int = -1,
-      clusteringColsOpt: Option[List[Column]] = None,
-      logCompactionInterval: Int = 10,
-      txnId: Option[(String, Long)] = None,
-      tablePropertiesRemoved: Set[String] = null): Transaction = {
-    // scalastyle:on argcount
-    val operation = if (isNewTable) Operation.CREATE_TABLE else Operation.WRITE
-
-    var txnBuilder = createWriteTxnBuilder(
-      TableImpl.forPath(engine, tablePath, clock),
-      operation)
-
-    if (isNewTable) {
-      txnBuilder = txnBuilder.withSchema(engine, schema)
-      if (partCols != null) {
-        txnBuilder = txnBuilder.withPartitionColumns(engine, partCols.asJava)
-      }
-    }
-
-    if (clusteringColsOpt.isDefined) {
-      txnBuilder = txnBuilder.withClusteringColumns(engine, clusteringColsOpt.get.asJava)
-    }
-
-    if (tableProperties != null) {
-      txnBuilder = txnBuilder.withTableProperties(engine, tableProperties.asJava)
-    }
-
-    if (withDomainMetadataSupported) {
-      txnBuilder = txnBuilder.withDomainMetadataSupported()
-    }
-
-    if (maxRetries >= 0) {
-      txnBuilder = txnBuilder.withMaxRetries(maxRetries)
-    }
-
-    txnBuilder = txnBuilder.withLogCompactionInverval(logCompactionInterval)
-
-    txnId.foreach { case (appId, txnVer) =>
-      txnBuilder = txnBuilder.withTransactionId(engine, appId, txnVer)
-    }
-    if (tablePropertiesRemoved != null) {
-      txnBuilder = txnBuilder.withTablePropertiesRemoved(tablePropertiesRemoved.asJava)
-    }
-    txnBuilder.build(engine)
-  }
-
-  override def createWriteTxnBuilder(
-      table: Table,
-      operation: Operation = Operation.WRITE): TransactionBuilder = {
-    table.createTransactionBuilder(defaultEngine, "test-engine", operation)
-  }
-}
-
+/** Default write utilities that use the V1 transaction builders. */
 trait WriteUtils extends AbstractWriteUtils with TransactionBuilderV1Support
 
 /**
- * Common utility methods for write test suites.
+ * Common utility methods for write test suites. For now, this includes mostly concrete
+ * implementations for the utilities. As we improve our test structure, we should move concrete
+ * implementations out of this class (like we have done for [[TransactionBuilderSupport]]). For
+ * example, `commitTransaction` could go into a [[TransactionCommitSupport]] trait since it is
+ * overridden in child suites.
  */
 trait AbstractWriteUtils extends TestUtils with TransactionBuilderSupport {
   val OBJ_MAPPER = new ObjectMapper()
